+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased] - Next release 0.101.3 or 1.0.0
 
 ### Added
+- Events generated when ingress is disabled due to content storage depletion
 
 ### Changed
 - Improved event summary for plugin installation/uninstall